First Coded Portfolio UX 220
Here is my first ever coded UX Portfolio for an assignment I did as part of my web design class. I mainly used HTML and CSS to complete the portfolio. I plan on refining my skills and continuing to learn more about coding to be able to create a more sophisticated portfolio in the future.